Add source sampling to amalgamated pyne
