+++ src/mcnp/CMakeLists.txt
@@ -16,10 +16,6 @@ message(STATUS "Building object library: mcnp_funcs")
 add_library(mcnp_funcs OBJECT mcnp_funcs.cpp)
 message(STATUS "Building object library: meshtal_funcs")
 add_library(meshtal_funcs OBJECT meshtal_funcs.cpp)
-message(STATUS "Building object library: measure")
-add_library(measure OBJECT measure.cpp)
-message(STATUS "Building object library: source_sampling")
-add_library(source_sampling OBJECT source_sampling.cpp)
 
 if (BUILD_MCNP5)
   add_subdirectory(mcnp5)
