+++ src/pyne/pyne.cpp
@@ -12,6 +12,8 @@
 //   src/material.cpp
 //   src/tally.cpp
 //   src/atomic_data.cpp
+//   src/measure.cpp
+//   src/source_sampling.cpp
 
 // PyNE amalgated source http://pyne.io/
 #include "pyne.h"
@@ -20133,6 +20135,788 @@ void pyne::_insert_abund_map() {
 }
 //
 // end of src/atomic_data.cpp
+//
+
+
+//
+// start of src/measure.cpp
+//
+/*
+ * Copyright (c) 2005 Lawrence Livermore National Laboratory under
+ * contract number B545069 with the University of Wisconsin - Madison.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <math.h>
+
+#include "measure.h"
+
+class CartVect {
+ private:
+  double coords[3];
+
+ public:
+
+  inline CartVect() {}
+
+  inline CartVect(double tx, double ty, double tz) { set(tx, ty, tz); }
+
+  inline CartVect(const CartVect& other) { set(other.coords); }
+
+  inline void set(double tx, double ty, double tz)
+  { coords[0] = tx; coords[1] = ty; coords[2] = tz; }
+
+  inline void set(const double* c)
+  { coords[0] = c[0]; coords[1] = c[1]; coords[2] = c[2]; }
+
+  inline double x() const { return coords[0]; }
+  inline double y() const { return coords[1]; }
+  inline double z() const { return coords[2]; }
+
+  inline CartVect& operator+=(const CartVect& other) {
+    coords[0] += other.coords[0];
+    coords[1] += other.coords[1];
+    coords[2] += other.coords[2];
+    return *this;
+  }
+
+  inline CartVect& operator-=(const CartVect& other) {
+    coords[0] -= other.coords[0];
+    coords[1] -= other.coords[1];
+    coords[2] -= other.coords[2];
+    return *this;
+  }
+
+  inline CartVect& operator*=(const CartVect& other);
+
+  inline double lensqr() const;
+
+  inline double len() const;
+
+  inline CartVect operator~() const;
+
+
+  inline CartVect& operator*=(double a) {
+    coords[0] *= a;
+    coords[1] *= a;
+    coords[2] *= a;
+    return *this;
+  }
+
+  inline CartVect& operator/=(double a) {
+    coords[0] /= a;
+    coords[1] /= a;
+    coords[2] /= a;
+    return *this;
+  }
+
+
+};
+
+inline CartVect operator+(const CartVect& v1, const CartVect& v2) {
+  CartVect rval(v1);
+  rval += v2;
+  return rval;
+}
+
+inline CartVect operator-(const CartVect& v1, const CartVect& v2) {
+  CartVect rval(v1);
+  rval -= v2;
+  return rval;
+}
+
+inline double operator%(const CartVect& v1, const CartVect& v2) {
+  return v1.x() * v2.x() + v1.y() * v2.y() + v1.z() * v2.z();
+}
+
+inline CartVect operator*(const CartVect& v1, const CartVect& v2) {
+  return CartVect(v1.y() * v2.z() - v1.z() * v2.y(),
+                  v1.z() * v2.x() - v1.x() * v2.z(),
+                  v1.x() * v2.y() - v1.y() * v2.x());
+}
+
+inline CartVect CartVect::operator~() const {
+  double invlen = 1.0 / len();
+  return CartVect(invlen * x(), invlen * y(), invlen * z());
+}
+
+inline CartVect& CartVect::operator*=(const CartVect& other)
+{ return *this = *this * other; }
+
+inline double CartVect::lensqr() const
+{ return *this % *this; }
+
+inline double CartVect::len() const
+{ return sqrt(lensqr()); }
+
+inline static double tet_volume(const CartVect& v0,
+                                const CartVect& v1,
+                                const CartVect& v2,
+                                const CartVect& v3) {
+  return 1. / 6. * (((v1 - v0) * (v2 - v0)) % (v3 - v0));
+}
+
+double edge_length(const double* start_vtx_coords,
+                   const double*   end_vtx_coords) {
+  const CartVect* start = reinterpret_cast<const CartVect*>(start_vtx_coords);
+  const CartVect*   end = reinterpret_cast<const CartVect*>(end_vtx_coords);
+  return (*start - *end).len();
+}
+
+double measure(moab::EntityType type,
+               int num_vertices,
+               const double* vertex_coordinates) {
+  const CartVect* coords = reinterpret_cast<const CartVect*>(vertex_coordinates);
+  switch (type) {
+    case moab::MBEDGE:
+      return (coords[0] - coords[1]).len();
+    case moab::MBTRI:
+      return 0.5 * ((coords[1] - coords[0]) * (coords[2] - coords[0])).len();
+    case moab::MBQUAD:
+      num_vertices = 4;
+    case moab::MBPOLYGON: {
+      CartVect mid(0, 0, 0);
+      for (int i = 0; i < num_vertices; ++i)
+        mid += coords[i];
+      mid /= num_vertices;
+
+      double sum = 0.0;
+      for (int i = 0; i < num_vertices; ++i) {
+        int j = (i + 1) % num_vertices;
+        sum += ((mid - coords[i]) * (mid - coords[j])).len();
+      }
+      return 0.5 * sum;
+    }
+    case moab::MBTET:
+      return tet_volume(coords[0], coords[1], coords[2], coords[3]) ;
+    case moab::MBPYRAMID:
+      return tet_volume(coords[0], coords[1], coords[2], coords[4]) +
+             tet_volume(coords[0], coords[2], coords[3], coords[4]) ;
+    case moab::MBPRISM:
+      return tet_volume(coords[0], coords[1], coords[2], coords[5]) +
+             tet_volume(coords[3], coords[5], coords[4], coords[0]) +
+             tet_volume(coords[1], coords[4], coords[5], coords[0]) ;
+    case moab::MBHEX:
+      return tet_volume(coords[0], coords[1], coords[3], coords[4]) +
+             tet_volume(coords[7], coords[3], coords[6], coords[4]) +
+             tet_volume(coords[4], coords[5], coords[1], coords[6]) +
+             tet_volume(coords[1], coords[6], coords[3], coords[4]) +
+             tet_volume(coords[2], coords[6], coords[3], coords[1]) ;
+    default:
+      return 0.0;
+  }
+}
+
+//
+// end of src/measure.cpp
+//
+
+
+//
+// start of src/source_sampling.cpp
+//
+#ifndef PYNE_IS_AMALGAMATED
+#include "source_sampling.h"
+#endif
+
+// Global sampler instance
+static pyne::Sampler* sampler = NULL;
+
+// Fortran API
+void pyne::sampling_setup_(int* mode) {
+  if (sampler == NULL) {
+    std::string filename("source.h5m");
+    std::string src_tag_name("source_density");
+    std::string e_bounds_file("e_bounds");
+    std::vector<double> e_bounds = read_e_bounds(e_bounds_file);
+    std::map<std::string, std::string> tag_names;
+    tag_names.insert(std::pair<std::string, std::string> ("src_tag_name",
+                                                          "source_density"));
+    tag_names.insert(std::pair<std::string, std::string> ("bias_tag_name",
+                                                          "biased_source_density"));
+    tag_names.insert(std::pair<std::string, std::string> ("cell_number_tag_name",
+                                                          "cell_number"));
+    tag_names.insert(std::pair<std::string, std::string> ("cell_fracs_tag_name",
+                                                          "cell_fracs"));
+    if (*mode == 0) {
+      sampler = new pyne::Sampler(filename, src_tag_name, e_bounds, false);
+    } else if (*mode == 1) {
+      sampler = new pyne::Sampler(filename, src_tag_name, e_bounds, true);
+    } else if (*mode == 2) {
+      std::string bias_tag_name("biased_source_density");
+      sampler = new pyne::Sampler(filename, src_tag_name, e_bounds, bias_tag_name);
+    } else {
+      sampler = new pyne::Sampler(filename, tag_names, e_bounds, *mode);
+    }
+  }
+}
+
+void pyne::particle_birth_(double* rands,
+                           double* x,
+                           double* y,
+                           double* z,
+                           double* e,
+                           double* w,
+                           int* c) {
+  std::vector<double> rands2(rands, rands + 6);
+  pyne::SourceParticle src = sampler->particle_birth(rands2);
+  *x = src.get_x();
+  *y = src.get_y();
+  *z = src.get_z();
+  *e = src.get_e();
+  *w = src.get_w();
+  *c = src.get_c();
+}
+
+std::vector<double> pyne::read_e_bounds(std::string e_bounds_file) {
+  std::vector<double> e_bounds;
+  std::ifstream inputFile(e_bounds_file.c_str());
+  double value;
+  if (inputFile) {
+    while (inputFile >> value)
+      e_bounds.push_back(value);
+  }
+  return e_bounds;
+}
+
+
+// C++ API
+pyne::Sampler::Sampler(std::string filename,
+                       std::string src_tag_name,
+                       std::vector<double> e_bounds,
+                       bool uniform)
+  : filename(filename), src_tag_name(src_tag_name), e_bounds(e_bounds) {
+  bias_mode = (uniform) ? UNIFORM : ANALOG;
+  sub_mode = DEFAULT;
+  setup();
+}
+
+pyne::Sampler::Sampler(std::string filename,
+                       std::string src_tag_name,
+                       std::vector<double> e_bounds,
+                       std::string bias_tag_name)
+  : filename(filename),
+    src_tag_name(src_tag_name),
+    e_bounds(e_bounds),
+    bias_tag_name(bias_tag_name) {
+  bias_mode = USER;
+  sub_mode = DEFAULT;
+  setup();
+}
+
+pyne::Sampler::Sampler(std::string filename,
+                       std::map<std::string, std::string> tag_names,
+                       std::vector<double> e_bounds,
+                       int mode)
+  : filename(filename),
+    e_bounds(e_bounds) {
+  // determine the bias_mode and sub_mode
+  if (mode == 0) {
+    bias_mode = ANALOG;
+    sub_mode = DEFAULT;
+  } else if (mode == 1) {
+    bias_mode = UNIFORM;
+    sub_mode = DEFAULT;
+  } else if (mode == 2) {
+    bias_mode = USER;
+    sub_mode = DEFAULT;
+  } else if (mode == 3) {
+    bias_mode = ANALOG;
+    sub_mode = SUBVOXEL;
+  } else if (mode == 4) {
+    bias_mode = UNIFORM;
+    sub_mode = SUBVOXEL;
+  } else if (mode == 5) {
+    bias_mode = USER;
+    sub_mode = SUBVOXEL;
+  }
+
+  // find out the src_tag_name and bias_tag_name
+  if (tag_names.find("src_tag_name") == tag_names.end()) {
+    // src_tag_name not found
+    throw std::invalid_argument("src_tag_name not found");
+  } else {
+    // found src_tag_name
+    src_tag_name = tag_names["src_tag_name"];
+  }
+  if (bias_mode == USER) {
+    // bias_tag_name required
+    if (tag_names.find("bias_tag_name") == tag_names.end()) {
+      // bias_tag_name not found
+      throw std::invalid_argument("bias_tag_name not found");
+    } else {
+      // found bias_tag_name
+      bias_tag_name = tag_names["bias_tag_name"];
+    }
+  }
+  if (sub_mode == SUBVOXEL) {
+    // cell_number_tag
+    if (tag_names.find("cell_number_tag_name") == tag_names.end()) {
+      // cell_number_tag_name not found
+      throw std::invalid_argument("cell_number_tag_name not found");
+    } else {
+      // found cell_number_tag_name
+      cell_number_tag_name = tag_names["cell_number_tag_name"];
+    }
+    // cell_fracs_tag
+    if (tag_names.find("cell_fracs_tag_name") == tag_names.end()) {
+      // cell_fracs_tag_name not found
+      throw std::invalid_argument("cell_fracs_tag_name not found");
+    } else {
+      // found cell_fracs_tag_name
+      cell_fracs_tag_name = tag_names["cell_fracs_tag_name"];
+    }
+  }
+  setup();
+}
+
+pyne::SourceParticle pyne::Sampler::particle_birth(std::vector<double> rands) {
+  // select mesh volume and energy group
+  // In DEFAULT mode, max_num_cells = 1
+  int pdf_idx = at->sample_pdf(rands[0], rands[1]);
+  int ve_idx = pdf_idx / max_num_cells / num_e_groups;
+  int c_idx = (pdf_idx / num_e_groups) % max_num_cells;
+  int e_idx = pdf_idx % num_e_groups;
+  int cell_id;
+
+  // Sample uniformly within the selected mesh volume element and energy
+  // group.
+  std::vector<double> xyz_rands;
+  xyz_rands.push_back(rands[2]);
+  xyz_rands.push_back(rands[3]);
+  xyz_rands.push_back(rands[4]);
+  moab::CartVect pos = sample_xyz(ve_idx, xyz_rands);
+  // cell_number
+  if (sub_mode == SUBVOXEL) {
+    cell_id = cell_number[ve_idx * max_num_cells + c_idx];
+  } else {
+    cell_id = -1;
+  }
+  pyne::SourceParticle src = SourceParticle(pos[0], pos[1], pos[2],
+                                            sample_e(e_idx, rands[5]), sample_w(pdf_idx), cell_id);
+  return src;
+}
+
+void pyne::Sampler::setup() {
+  moab::ErrorCode rval;
+  moab::EntityHandle loaded_file_set;
+  // Create MOAB instance
+  mesh = new moab::Core();
+  rval = mesh->create_meshset(moab::MESHSET_SET, loaded_file_set);
+  rval = mesh->load_file(filename.c_str(), &loaded_file_set);
+  if (rval != moab::MB_SUCCESS)
+    throw std::invalid_argument("Could not load mesh file.");
+
+  // Get mesh volume elemebts
+  moab::Range ves;
+  rval = mesh->get_entities_by_dimension(loaded_file_set, 3, ves);
+  if (rval != moab::MB_SUCCESS)
+    throw std::runtime_error("Problem entities of dimension 3");
+  num_ves = ves.size();
+  int num_hex, num_tet;
+  rval = mesh->get_number_entities_by_type(loaded_file_set, moab::MBHEX, num_hex);
+  rval = mesh->get_number_entities_by_type(loaded_file_set, moab::MBTET, num_tet);
+  if (num_hex == num_ves) {
+    ve_type = moab::MBHEX;
+    verts_per_ve = 8;
+  } else if (num_tet == num_ves) {
+    ve_type = moab::MBTET;
+    verts_per_ve = 4;
+  } else
+    throw std::invalid_argument("Mesh file must contain only tets or hexes.");
+
+  // Process all the spatial and tag data and create an alias table.
+  std::vector<double> volumes(num_ves);
+  mesh_geom_data(ves, volumes);
+  mesh_tag_data(ves, volumes);
+}
+
+void pyne::Sampler::mesh_geom_data(moab::Range ves, std::vector<double>& volumes) {
+  // Get connectivity.
+  moab::ErrorCode rval;
+  std::vector<moab::EntityHandle> connect;
+  rval = mesh->get_connectivity_by_type(ve_type, connect);
+  if (rval != moab::MB_SUCCESS)
+    throw std::runtime_error("Problem getting mesh connectivity.");
+
+  // Grab the coordinates that define 4 connected points within a mesh volume
+  // element and setup a data structure to allow uniform sampling with each
+  // mesh volume element.
+  double coords[verts_per_ve * 3];
+  int v;
+  for (v = 0; v < num_ves; ++v) {
+    rval = mesh->get_coords(&connect[verts_per_ve * v], verts_per_ve, &coords[0]);
+    if (rval != moab::MB_SUCCESS)
+      throw std::runtime_error("Problem vertex coordinates.");
+    volumes[v] = measure(ve_type, verts_per_ve, &coords[0]);
+    if (ve_type == moab::MBHEX) {
+      moab::CartVect o(coords[0], coords[1], coords[2]);
+      moab::CartVect x(coords[3], coords[4], coords[5]);
+      moab::CartVect y(coords[9], coords[10], coords[11]);
+      moab::CartVect z(coords[12], coords[13], coords[14]);
+      edge_points ep = {o, x - o, y - o, z - o};
+      all_edge_points.push_back(ep);
+    } else if (ve_type == moab::MBTET) {
+      moab::CartVect o(coords[0], coords[1], coords[2]);
+      moab::CartVect x(coords[3], coords[4], coords[5]);
+      moab::CartVect y(coords[6], coords[7], coords[8]);
+      moab::CartVect z(coords[9], coords[10], coords[11]);
+      edge_points ep = {o, x - o, y - o, z - o};
+      all_edge_points.push_back(ep);
+    }
+  }
+}
+
+void pyne::Sampler::mesh_tag_data(moab::Range ves,
+                                  const std::vector<double> volumes) {
+  moab::ErrorCode rval;
+  moab::Tag src_tag;
+  moab::Tag cell_number_tag;
+  moab::Tag cell_fracs_tag;
+  rval = mesh->tag_get_handle(src_tag_name.c_str(),
+                              moab::MB_TAG_VARLEN,
+                              moab::MB_TYPE_DOUBLE,
+                              src_tag);
+  // THIS rval FAILS because we do not know number of energy groups a priori.
+  // That's okay. That's what the next line is all about:
+  num_e_groups = num_groups(src_tag);
+
+  // Set the default value of max_num_cells to 1, so that the normal r2s and sub-voxel
+  // r2s can use the same form of pdf size description
+  max_num_cells = 1;
+  // set the default value of cell_fracs to 1.0
+  cell_fracs.resize(num_ves * max_num_cells);
+  for (int i = 0; i < cell_fracs.size(); i++) {
+    cell_fracs[i] = 1.0;
+  }
+  if (sub_mode == SUBVOXEL) {
+    // Read the cell_number tag and cell_fracs tag
+    rval = mesh->tag_get_handle(cell_number_tag_name.c_str(),
+                                cell_number_tag);
+    rval = mesh->tag_get_handle(cell_fracs_tag_name.c_str(),
+                                cell_fracs_tag);
+    max_num_cells = num_groups(cell_fracs_tag);
+    num_e_groups /= max_num_cells;
+    cell_fracs.resize(num_ves * max_num_cells);
+    rval = mesh->tag_get_data(cell_fracs_tag, ves, &cell_fracs[0]);
+    cell_number.resize(num_ves * max_num_cells);
+    rval = mesh->tag_get_data(cell_number_tag, ves, &cell_number[0]);
+  }
+  std::vector<double> pdf(num_ves * num_e_groups * max_num_cells);
+  rval = mesh->tag_get_data(src_tag, ves, &pdf[0]);
+  if (rval != moab::MB_SUCCESS)
+    throw std::runtime_error("Problem getting source tag data.");
+
+  // Multiply the source densities by the VE volumes
+  int v, c, e;
+  for (v = 0; v < num_ves; ++v) {
+    for (c = 0; c < max_num_cells; ++c) {
+      for (e = 0; e < num_e_groups; ++e) {
+        pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] *=
+            volumes[v] * cell_fracs[v * max_num_cells + c];
+      }
+    }
+  }
+  normalize_pdf(pdf);
+
+  // Setup alias table based off PDF or biased PDF
+  if (bias_mode == ANALOG) {
+    at = new AliasTable(pdf);
+  } else {
+    std::vector<double> bias_pdf = read_bias_pdf(ves, volumes, pdf);
+    normalize_pdf(bias_pdf);
+    //  Create alias table based off biased pdf and calculate birth weights.
+    biased_weights.resize(num_ves * num_e_groups * max_num_cells);
+    for (int i = 0; i < biased_weights.size(); ++i) {
+      biased_weights[i] = pdf[i] / bias_pdf[i];
+    }
+    at = new AliasTable(bias_pdf);
+  }
+}
+
+std::vector<double> pyne::Sampler::read_bias_pdf(moab::Range ves,
+                                                 std::vector<double> volumes,
+                                                 std::vector<double> pdf) {
+  std::vector<double> bias_pdf(num_ves * max_num_cells * num_e_groups);
+  int v, c, e;
+  moab::ErrorCode rval;
+  if (bias_mode == UNIFORM) {
+    // Sub-voxel Uniform sampling: uniform in space, analog in energy. Biased PDF is
+    // found by normalizing the total photon emission density to 1 in each
+    // mesh volume element and multiplying by the volume of the element.
+    double q_in_group;
+    for (v = 0; v < num_ves; ++v) {
+      for (c = 0; c < max_num_cells; ++c) {
+        q_in_group = 0.0;
+        for (e = 0; e < num_e_groups; ++e) {
+          q_in_group += pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e];
+        }
+
+        if (q_in_group > 0) {
+          for (e = 0; e < num_e_groups; ++e) {
+            bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] =
+                volumes[v] * cell_fracs[v * max_num_cells + c] *
+                pdf[v * max_num_cells * num_e_groups +
+                      c * num_e_groups + e] / q_in_group;
+          }
+        } else {
+          for (e = 0; e < num_e_groups; ++e) {
+            bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] = 0.0;
+          }
+        }
+
+      }
+    }
+  } else if (bias_mode == USER) {
+    // Get the biased PDF from the mesh
+    moab::Tag bias_tag;
+    rval = mesh->tag_get_handle(bias_tag_name.c_str(),
+                                moab::MB_TAG_VARLEN,
+                                moab::MB_TYPE_DOUBLE,
+                                bias_tag);
+    num_bias_groups = num_groups(bias_tag);
+    if (num_bias_groups == num_e_groups * max_num_cells) {
+      // Spatial, cell and energy biasing. The supplied bias PDF values are
+      // applied to each specific energy group and sub-voxels in a mesh
+      // volume element.
+      rval = mesh->tag_get_data(bias_tag, ves, &bias_pdf[0]);
+      if (rval != moab::MB_SUCCESS)
+        throw std::runtime_error("Problem getting bias tag data.");
+      for (v = 0; v < num_ves; ++v) {
+        for (c = 0; c < max_num_cells; c++) {
+          for (e = 0; e < num_e_groups; ++e)
+            bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] *=
+                volumes[v] * cell_fracs[v * max_num_cells + c];
+        }
+      }
+    } else if (num_bias_groups == 1) {
+      // Spatial biasing only: the supplied bias PDF values are applied
+      // to all energy groups within a mesh volume element, which are
+      // sampled in analog.
+      std::vector<double> spatial_pdf(num_ves);
+      rval = mesh->tag_get_data(bias_tag, ves, &spatial_pdf[0]);
+      if (rval != moab::MB_SUCCESS)
+        throw std::runtime_error("Problem getting bias tag data.");
+      double q_in_group;
+      for (v = 0; v < num_ves; ++v) {
+        q_in_group = 0;
+        for (c = 0; c < max_num_cells; ++c) {
+          for (e = 0; e < num_e_groups; ++e) {
+            q_in_group += pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e];
+          }
+        }
+        if (q_in_group > 0) {
+          for (c = 0; c < max_num_cells; ++c) {
+            for (e = 0; e < num_e_groups; ++e) {
+              bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] =
+                  spatial_pdf[v] * volumes[v] * cell_fracs[v * max_num_cells + c] *
+                  pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] /
+                  q_in_group;
+            }
+          }
+        } else {
+          for (c = 0; c < max_num_cells; ++c)
+            for (e = 0; e < num_e_groups; ++e) {
+              bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] =  0;
+            }
+        }
+      }
+    } else if (num_bias_groups == num_e_groups) {
+      // Voxel and energy biasing. Apply the energy bias to all the sub-voxel in the voxel
+      std::vector<double> spa_erg_pdf(num_ves * num_e_groups);
+      rval = mesh->tag_get_data(bias_tag, ves, &spa_erg_pdf[0]);
+      if (rval != moab::MB_SUCCESS)
+        throw std::runtime_error("Problem getting bias tag data.");
+      double q_in_group;
+      for (v = 0; v < num_ves; ++v) {
+        for (e = 0; e < num_e_groups; ++e) {
+          q_in_group = 0.0;
+          for (c = 0; c < max_num_cells; ++c) {
+            q_in_group += pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e];
+          }
+          if (q_in_group > 0) {
+            for (c = 0; c < max_num_cells; ++c) {
+              bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] =
+                  spa_erg_pdf[v * num_e_groups + e] * volumes[v] * cell_fracs[v * max_num_cells + c] *
+                  pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] / q_in_group;
+            }
+          } else {
+            for (c = 0; c < max_num_cells; ++c) {
+              bias_pdf[v * max_num_cells * num_e_groups + c * num_e_groups + e] = 0.0;
+            }
+          }
+        }
+      }
+    } else {
+      throw std::length_error("Length of bias tag must equal length of the"
+                              "  max_num_cells*num_e_group, num_e_groups, or 1.");
+    }
+  }
+  double q_in_all = 0.0;
+  for (int i = 0; i < bias_pdf.size(); i++)
+    q_in_all += bias_pdf[i];
+  if (q_in_all <= 0.0) {
+    throw std::runtime_error("Bias data are ALL ZERO!");
+  }
+  return bias_pdf;
+}
+
+
+moab::CartVect pyne::Sampler::sample_xyz(int ve_idx, std::vector<double> rands) {
+  double s = rands[0];
+  double t = rands[1];
+  double u = rands[2];
+
+  // Transform s, t, u to uniformly sample a tetrahedron. See:
+  // C. Rocchini and P. Cignoni, “Generating Random Points in a Tetrahedron,”
+  //  Journal of Graphics Tools, 5, 200–202 (2001).
+  if (ve_type == moab::MBTET) {
+    if (s + t > 1) {
+      s = 1.0 - s;
+      t = 1.0 - t;
+    }
+    if (s + t + u > 1) {
+      if (t + u > 1) {
+        double old_t = t;
+        t = 1.0 - u;
+        u = 1.0 - s - old_t;
+      } else if (t + u <= 1) {
+        double old_s = s;
+        s = 1.0 - t - u;
+        u = old_s + t + u - 1;
+      }
+    }
+  }
+
+  return s * all_edge_points[ve_idx].x_vec + \
+         t * all_edge_points[ve_idx].y_vec + \
+         u * all_edge_points[ve_idx].z_vec + \
+         all_edge_points[ve_idx].o_point;
+}
+
+double pyne::Sampler::sample_e(int e_idx, double rand) {
+  double e_min = e_bounds[e_idx];
+  double e_max = e_bounds[e_idx + 1];
+  return rand * (e_max - e_min) + e_min;
+}
+
+double pyne::Sampler::sample_w(int pdf_idx) {
+  return (bias_mode == ANALOG) ? 1.0 : biased_weights[pdf_idx];
+}
+
+void pyne::Sampler::normalize_pdf(std::vector<double>& pdf) {
+  double sum = 0;
+  for (int i = 0; i < pdf.size(); ++i)
+    sum += pdf[i];
+  for (int i = 0; i < pdf.size(); ++i)
+    pdf[i] /= sum;
+}
+
+int pyne::Sampler::num_groups(moab::Tag tag) {
+  moab::ErrorCode rval;
+  int tag_size;
+  rval = mesh->tag_get_bytes(tag, *(&tag_size));
+  if (rval != moab::MB_SUCCESS)
+    throw std::runtime_error("Problem getting tag size.");
+  return tag_size / sizeof(double);
+}
+
+
+// Random-number sampling using the Walker-Vose alias method,
+// Copyright: Joachim Wuttke, Forschungszentrum Juelich GmbH (2013)
+// M. D. Vose, IEEE T. Software Eng. 17, 972 (1991)
+// A. J. Walker, Electronics Letters 10, 127 (1974); ACM TOMS 3, 253 (1977)
+
+pyne::AliasTable::AliasTable(std::vector<double> p) {
+  n = p.size();
+  prob.resize(n);
+  alias.resize(n);
+  std::vector<double> small(n);
+  std::vector<double> large(n);
+  int i, a, g;
+
+  for (i = 0; i < n; ++i)
+    p[i] *= n;
+
+  // Set separate index lists for small and large probabilities:
+  int n_s = 0;
+  int n_l = 0;
+  for (i = n - 1; i >= 0; --i) {
+    // at variance from Schwarz, we revert the index order
+    if (p[i] < 1)
+      small[n_s++] = i;
+    else
+      large[n_l++] = i;
+  }
+
+  // Work through index lists
+  while (n_s && n_l) {
+    a = small[--n_s]; // Schwarz's l
+    g = large[--n_l]; // Schwarz's g
+    prob[a] = p[a];
+    alias[a] = g;
+    p[g] = p[g] + p[a] - 1;
+    if (p[g] < 1)
+      small[n_s++] = g;
+    else
+      large[n_l++] = g;
+  }
+
+  while (n_l)
+    prob[large[--n_l]] = 1;
+
+  while (n_s)
+    // can only happen through numeric instability
+    prob[small[--n_s] ] = 1;
+}
+
+int pyne::AliasTable::sample_pdf(double rand1, double rand2) {
+  int i = (int) n * rand1;
+  return rand2 < prob[i] ? i : alias[i];
+}
+
+pyne::SourceParticle::SourceParticle() {
+  x = -1.0;
+  y = -1.0;
+  z = -1.0;
+  e = -1.0;
+  w = -1.0;
+  c = -1;
+}
+
+pyne::SourceParticle::SourceParticle(double _x, double _y, double _z,
+                                     double _e, double _w, int _c) {
+  x = _x;
+  y = _y;
+  z = _z;
+  e = _e;
+  w = _w;
+  c = _c;
+}
+
+pyne::SourceParticle::~SourceParticle() {};
+
+//
+// end of src/source_sampling.cpp
 //
