+++ src/pyne/update_pyne.sh
@@ -9,7 +9,8 @@ cd ..
 python amalgamate.py -f license.txt src/utils.* src/extra_types.h src/h5wrap.h \
     src/state_map.cpp src/nucname.* src/rxname.* src/particle.* src/data.* \
     src/json-forwards.h src/json.h src/jsoncpp.cpp src/jsoncustomwriter.* \
-    src/material.* src/tally.* src/atomic_data.*
+    src/material.* src/tally.* src/atomic_data.* src/measure.* \
+    src/source_sampling.*
 cp pyne.cpp pyne.h ..
 cd ..
 rm -rf pyne
